Add multi-value progress bar for goals
This will help display the proportion of the goal's progress that is safe (That is, is not a critical component of a loan to a purchased goal), unsafe (Propertion that has been accrued but is a critical component of a loan), available (That can be loaned to it from other goals), and remaining. Different styling is offered for purchased goals to mark safe and remainder in a more prominent way.